Starting to rip out regular expression code for refs / paths
