+++ go/libraries/doltcore/doltdb/doltdb.go
@@ -216,7 +216,7 @@ func (ddb *DoltDB) WriteEmptyRepoWithCommitTimeAndDefaultBranch(
 }
 
 func getCommitValForRefStr(ctx context.Context, db datas.Database, vrw types.ValueReadWriter, ref string) (*datas.Commit, error) {
-	if !datas.DatasetFullRe.MatchString(ref) {
+	if err := datas.ValidateDatasetId(ref); err != nil {
 		return nil, fmt.Errorf("invalid ref format: %s", ref)
 	}
