+++ go/store/cmd/noms/noms_log_test.go
@@ -28,7 +28,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/dolthub/dolt/go/store/d"
 	"github.com/dolthub/dolt/go/store/datas"
 	"github.com/dolthub/dolt/go/store/hash"
 	"github.com/dolthub/dolt/go/store/spec"
@@ -125,23 +124,6 @@ func mustHead(ds datas.Dataset) types.Value {
 		panic("no head")
 	}
 	return s
-}
-
-func mustHeadAddr(ds datas.Dataset) hash.Hash {
-	addr, ok := ds.MaybeHeadAddr()
-	d.PanicIfFalse(ok)
-	return addr
-}
-
-func mustHeadValue(ds datas.Dataset) types.Value {
-	val, ok, err := ds.MaybeHeadValue()
-	d.PanicIfError(err)
-
-	if !ok {
-		panic("no head")
-	}
-
-	return val
 }
 
 func (s *nomsLogTestSuite) TestNArg() {
