+++ go/store/cmd/noms/noms_merge_test.go
