+++ go/store/cmd/noms/noms_show.go
@@ -24,9 +24,11 @@ package main
 import (
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"io"
 	"os"
+	"time"
 
 	flag "github.com/juju/gnuflag"
 
@@ -209,5 +211,18 @@ func outputEncodedValue(ctx context.Context, w io.Writer, value types.Value) err
 		}
 	default:
 		return types.WriteEncodedValue(ctx, w, value)
+	}
+}
+
+func locationFromTimezoneArg(tz string, defaultTZ *time.Location) (*time.Location, error) {
+	switch tz {
+	case "local":
+		return time.Local, nil
+	case "utc":
+		return time.UTC, nil
+	case "":
+		return defaultTZ, nil
+	default:
+		return nil, errors.New("value must be: local or utc")
 	}
 }
