+++ go/store/cmd/noms/noms_sync_test.go
@@ -25,6 +25,8 @@ import (
 	"context"
 	"testing"
 
+	"github.com/dolthub/dolt/go/store/d"
+	"github.com/dolthub/dolt/go/store/hash"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/dolthub/dolt/go/libraries/utils/file"
@@ -207,4 +209,21 @@ func (s *nomsSyncTestSuite) TestRewind() {
 	s.NoError(err)
 	s.True(types.Float(42).Equals(mustHeadValue(dest)))
 	db.Close()
+}
+
+func mustHeadValue(ds datas.Dataset) types.Value {
+	val, ok, err := ds.MaybeHeadValue()
+	d.PanicIfError(err)
+
+	if !ok {
+		panic("no head")
+	}
+
+	return val
+}
+
+func mustHeadAddr(ds datas.Dataset) hash.Hash {
+	addr, ok := ds.MaybeHeadAddr()
+	d.PanicIfFalse(ok)
+	return addr
 }
